chore: rename UtxoDetailModule -> UtxoDetailModal
