+++ src/components/jar_details/JarDetailsOverlay.tsx
@@ -11,7 +11,7 @@ import Alert, { SimpleMessageAlertProps } from '../Alert'
 import Balance from '../Balance'
 import Sprite from '../Sprite'
 import SegmentedTabs from '../SegmentedTabs'
-import UtxoDetailModal from './UtxoDetailModule'
+import UtxoDetailModal from './UtxoDetailModal'
 import { UtxoList } from './UtxoList'
 import { DisplayBranchHeader, DisplayBranchBody } from './DisplayBranch'
 import { jarInitial, jarName } from '../jars/Jar'
