+++ src/components/jar_details/UtxoDetailModal.module.css
